refactor(ai-provider,shared): 适配器去重、模型数据外置、按组拆分
- adapters/common.ts 统一参数提取/元数据清理/mediaRefFallbackText
- known-models 950→66，数据外置为 known-models.data.json
- gateway fallback 去重(gateway-fallback-chain.ts)
- env registry 651→51 按 8 组拆分；plugin schema 528→12 拆 schema/types
- 对外公共 API 零变更，ai-provider 163 + shared 56 测试全过
