+++ packages/ai-provider/src/adapters/openai-chat.ts
@@ -19,8 +19,13 @@ import {
   readOpenAiChatStreamToolCallDeltas,
 } from "./http.js";
 import { applyCapabilityFallback } from "./capability-fallback.js";
+import {
+  createMetadataSanitizer,
+  extractParameterOverrides,
+  mediaRefFallbackText,
+} from "./common.js";
 
-import type { ImagePart, TextMessage, TextMessageContent } from "../types.js";
+import type { TextMessage, TextMessageContent } from "../types.js";
 
 /** Fields that providerRequestMetadata must never override. */
 const OPENAI_PROTECTED_KEYS = new Set([
@@ -37,38 +42,21 @@ const OPENAI_PROTECTED_KEYS = new Set([
   "parameterOverrides",
 ]);
 
-function sanitizeOpenAiMetadata(
-  meta: Record<string, unknown> | undefined,
-): Record<string, unknown> {
-  if (!meta) return {};
-  const sanitized: Record<string, unknown> = {};
-  for (const [k, v] of Object.entries(meta)) {
-    if (!OPENAI_PROTECTED_KEYS.has(k)) sanitized[k] = v;
-  }
-  return sanitized;
-}
+/** camelCase override key → OpenAI Chat wire field. */
+const OPENAI_PARAMETER_FIELD_MAP = {
+  temperature: "temperature",
+  topP: "top_p",
+  maxOutputTokens: "max_tokens",
+  frequencyPenalty: "frequency_penalty",
+  presencePenalty: "presence_penalty",
+} as const;
+
+const sanitizeOpenAiMetadata = createMetadataSanitizer(OPENAI_PROTECTED_KEYS);
 
 function extractOpenAiParameterOverrides(
   meta: Record<string, unknown> | undefined,
 ): Record<string, unknown> {
-  const overrides = meta?.parameterOverrides;
-  if (!overrides || typeof overrides !== "object" || Array.isArray(overrides)) {
-    return {};
-  }
-  const source = overrides as Record<string, unknown>;
-  const result: Record<string, unknown> = {};
-  if (typeof source.temperature === "number")
-    result.temperature = source.temperature;
-  if (typeof source.topP === "number") result.top_p = source.topP;
-  if (typeof source.maxOutputTokens === "number")
-    result.max_tokens = source.maxOutputTokens;
-  if (typeof source.frequencyPenalty === "number") {
-    result.frequency_penalty = source.frequencyPenalty;
-  }
-  if (typeof source.presencePenalty === "number") {
-    result.presence_penalty = source.presencePenalty;
-  }
-  return result;
+  return extractParameterOverrides(meta, OPENAI_PARAMETER_FIELD_MAP);
 }
 
 /**
@@ -87,14 +75,6 @@ function buildEmbeddingInput(values: string[], format: string): unknown {
     }));
   }
   return values;
-}
-
-function mediaRefFallbackText(part: ImagePart): string {
-  return JSON.stringify({
-    type: "image_ref",
-    ref: part.image,
-    note: "MediaRef has no resolved URL; provider vision input requires a retrievable image URL.",
-  });
 }
 
 function serializeOpenAiChatContent(content: TextMessageContent): unknown {
